requirements.txt
asgiref==3.8.1
Django==5.2.3
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pytesseract==0.3.13
sqlparse==0.5.3
typing_extensions==4.14.0
